src/main/mowitnow/tondeuse/entities/Position.java
package main.mowitnow.tondeuse.entities;

public class Position {

    private int x;
    private int y;

    public int getX() {
        return x;
    }

    public int getY() {
        return y;
    }

    public Position(int x, int y) {
        this.x = x;
        this.y = y;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Position position = (Position) o;
        return x == position.x && y == position.y;
    }

    @Override
    public int hashCode() {
        final int prime = 31;
        int result = 1;
        result = prime * result + x;
        result = prime * result + y;
        return result;
    }

    public boolean isPositionValid(Position position) {
        return position.getX() >= 0
                && position.getY() >= 0
                && position.getX() <= this.x
                && position.getY() <= this.y;
    }
}
